agents/requirements.txt
openai>=1.40.0
pydantic>=2.7.0
pyyaml>=6.0.2
